Improve default dialyzer PLT by trying to guess dependencies
Try to automatically detect all application dependencies when `plt_apps`
is not included in `rebar.config`.

Note that this will not follow `runtime_dependencies` in OTP
applications. This can be resolved by adding any missing
`runtime_dependencies` to `applications` in the .app.src file or
including `plt_apps` in `rebar.config`.