Add packages to code path after installation
Fix a bug where packages are not added to the code path after
installation. Dependent applications that build from source are not
affected by this issue since the build_apps function in
rebar_prv_compiler takes care of the code path changes for them. It is
only the precompiled packages that suffer from this issue.