Preserve erl_opts from parent state in new state
Preserve the erl_opts set in the parent state when creating a new state
with rebar_state:new/3. This allows options that are set by intermediate
providers or plugins to persist through a command execution. This is
useful for example when creating an eunit testing plugin that passes
compiler options to be used when building files for test
execution. Without this change those options are lost when the eunit
provider creates a new state to pass to rebar_erlc_compiler.