Attempt at safe profile reapplication
Because opts are being reused and re-merged, reapplying profiles is done
by a base mechanism:

- Base opts are stored as-is
- Added opts are stored separately
- There is a merged opts used for quick reads/overrides

All reads are done from the merged opts, all writes to both the merged
and added opts.

When profiles are merged, this is done from the original base opts, and
then the new added opts are reapplied.