Improve semantic versioning from git between tags
In the scenario we that we have selected a commit
that is between two tags, we should base the
version on the most recent tag we can see in the
revision history, but we should not treat this as
the tag version.

`git describe --tags --abbrev=0` finds the most
recent tag visible in the revision history from
the current HEAD. Return this as the version
string and undefined as the tag to trigger ref
counting.