generalize compile suite to run umbrella configs
The cases in rebar_compile_SUITE can now be run as a plain application
AND as an application in an umbrella project.  The config "apps" can
be thought of as the project directory, and a new config "subject_dir"
indicates where the test application is to be created.  In
"as_application" mode, "apps" and "subject_dir" are identical.
Existing cases are backwards compatible with "as_application" mode. In
"as_umbrella", "subject_dir" will be some subdirectory of "apps".

To be compatible with "as_umbrella", the test case must replace some
uses of "apps" with "subject_dir".  Some cases have been updated in
this patch to demonstrate this, but the main focus here is to
establish the dual-use infrastructure.