[xref] Add applications from .app file to library path
Only applications listed under 'applications' and 'included_applications'
in .app file are added to library path when doing xref analysis.
The result of this is a new warning when calling code that is in the path
of an application but the application is not directly depending upon.
Eg. Calling a module in the dependency of one of our dependencies will
give a warning.
The config {xref_include_all_code, true} will keep the old behaviour
where all coded in path is considered callable.