Plugin templates enabled
This lets a plugin define templates to be loaded:

    $ rebar3 new
    ...
    proper (plugin): A basic PropEr suite for an OTP application
    ...
    $ rebar3 new help proper
    proper:
            plugin template (...)
            Description: A basic PropEr suite for an OTP application
            Variables:
                    name="suite" (...)
                    ...
    → rebar3 new proper fakesuite
    ===> Writing test/prop_fakesuite.erl

In this case, proper is a fake template file I've put by hand in
_build/default/plugins/rebar3_proper/priv/<somename>/, meaning it will
only work as far as it's called from the project's root.

The priority order of plugins is now .config > plugin > built-in, such
that someone could ensure plugins do not crush their own private
templates, but also that custom or plugin templates do overtake built-in
ones. It used to be Built-in > .config only.

Templates are searched for recursively in the priv/ directory of
plugins.