Correct TTY type detection
This reuses the trick used within OTP to pick within old and new shell.

The 'user' structure is the same for all cases (escript, escript + dumb
TERM, unstable install, unstable install + dumb TERM), so we take it
down first.

Then we boot the TTY driver, which fails if TERM=dumb, in which case we
boot the retro-style usr. If it worked, we shut down the driver again,
and boot a modern shell structure.

This avoids all warnings and seems to work in all cases.