Escriptize based on configured apps only
Prior to this patch, the escriptize command flat out selected all
declared dependencies.

This patch instead looks at the app files and only includes the
dependencies of the top level app and the extra ones, avoiding to
package more apps than required.

This required a version bump on cth_readable as it mistakenly included
'syntax_lib' instead of 'syntax_tools' as a dependency.