Track package hash in memory index, add hash test
This adds tracking of package hash in the in-memory index rather than
the current `undefined' values.

According to the test added, this is not necessary for transitive
package dep hash chcking, but does result in a more complete index search
result when doing app lookups, and could yield some optimizations on
hash checks by checking from the index structure before fetching a package.