Allow for override of builtin resources.
See https://github.com/erlang/rebar3/issues/1283 for some backstory.

This allows a plugin to replace a builtin resource and have it stick
while other plugins and providers are running.  Specifically when
compiling the resource might be used to check the version, but the path
to the resource might have been excised.  This adds a new function
called replace_resource to the rebar_state module.  When you replace a
resource it removes it's code path from the all_plugin_deps code path
and instead puts it into the all_deps code path.