Handle internal CT failures
This is based on issue #1517 where out of nowhere, CT has returned a
user's error code directly. This in turn caused a crashdump in rebar3
itself.

This patch handles the unexpected cases by:

a) not trying to format them
b) converting them to an error whenever they happen

The execution flow is still interrupted, but we should fail with a
clearer error than a crashdump.