Upgrade certifi dependencies
Get more up-to-date root certs. This was blocked for a while because
certifi wouldn't build on windows, but this is now fixed, with minimal
changes to the use case -- only a small change in the bootstrap script
is required.

The new certifi lib is also a few megabytes lighter than before, which
is good for rebar3.